Creation of Hybrid for Webhooks Transfer
Control and Worker nodes are split between on-premise and cloud
machines, and currently the on-premise nodes are behind dynamic IPs,
making the Webhooks service on the Control nodes useless.
This commit creates a new Hybrid node that runs as a Control & Worker
node. It also unifies Nginx reverse-proxies into one ingress, and
delegates the setting of the 'nginx_conf' var to the host instead of the
group.