Creation of Update-Hybrid Playbook
Updating a Hybrid node requires checking out the updated repo, i.e.
main, before running hybrid.yml.
This commit creates a new playbook specifically to update hybrid, this
is preferred to adding the 'git' role with a 'never' tag to the original
hybrid.yml. See the comment in update-hybrid.yml for further detail.